Move Lower Third settings tab back to composeApp
Relocates `LowerThirdSettingsTab`, its ViewModel, and related tests from `:lower-third-tab` into `:composeApp` so the options-dialog page stays with the rest of the dialog tabs. Updated package/import references (including `OptionsDialog`), renamed the shared test helper to `lowerThirdTab`, and moved the tab’s screenshot assets under `composeApp/screenshots`.
