+++ composeApp/screenshots/lowerThirdSettingsTab/band_60_percent.png
