+++ composeApp/screenshots/lowerThirdSettingsTab/empty_folder.png
