+++ composeApp/screenshots/lowerThirdSettingsTab/preset_selected.png
