+++ composeApp/screenshots/lowerThirdSettingsTab/preset_selected_other.png
