+++ composeApp/screenshots/lowerThirdSettingsTab/presets.png
