+++ composeApp/screenshots/lowerThirdSettingsTab/window_insets.png
